requirements.txt
boto3==1.28.2
notion-client==1.2.0
pyarrow==12.0.1
pyyaml==6.0
python-dotenv==1.0.0